Create separate files for classes in lib folder
