+++ bin/main.rb
@@ -1,5 +1,7 @@
 #!/usr/bin/env ruby
 require './lib/game.rb'
+require './lib/board.rb'
+require './lib/player.rb'
 
 def display_board(prepared_data, row_separator = '-------------')
   puts row_separator
