+++ lib/board.rb
@@ -1,0 +1,11 @@
+class Board
+  attr_accessor :b_values
+
+  def initialize(values)
+    @b_values = values
+  end
+
+  def reload
+    # logic to re-initialize the board values
+  end
+end
