+++ lib/game.rb
@@ -51,32 +51,3 @@ class Game
   # def process_move(input)
   # end
 end
-
-class Player
-  attr_accessor :name
-
-  def initialize(name)
-    @name = name
-  end
-
-  def move
-    # logic to move (play)
-  end
-
-  def false_move_message
-    "Invalid move by #{name}, try again!"
-  end
-
-end
-
-class Board
-  attr_accessor :b_values
-
-  def initialize(values)
-    @b_values = values
-  end
-
-  def reload
-    # logic to re-initialize the board values
-  end
-end
